requirements.txt
yandex-cloud-ml-sdk>=0.17.0
python-telegram-bot>=20.0

